requirements.txt
# Core dependencies - Latest AutoGen 0.4
pyautogen>=0.4.0
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Vector Database and Embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-openai>=0.1.0

# Traditional Database (for structured data)
sqlalchemy>=2.0.0
alembic>=1.12.0

# WhatsApp Integration (Primary Focus)
whatsapp-web.js>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pywhatkit>=5.4.0

# Social media APIs (Configurable/Optional)
tweepy>=4.14.0
instaloader>=4.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
textblob>=0.17.1
vaderSentiment>=3.3.2
transformers>=4.35.0
torch>=2.1.0

# LLM Integration
openai>=1.12.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0
python-dateutil>=2.8.2
pytz>=2023.3
asyncio-mqtt>=0.13.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Optional: For advanced NLP
spacy>=3.7.0
nltk>=3.8.1

# Documentation generation
markdown>=3.5.0
jinja2>=3.1.0
